requirements.txt
Django==1.6.6
psycopg2
dj-database-url
South
python-social-auth
djangorestframework